PDDL generation frame made
The base frame has been made, what it will be filled with still needs to
be done. Feel free to play around with it by running the "AgentTest"
scene and clicking on MasterAgent and clicking "Make File". It will make
one according to the given name in the "test name".