Writing Objects and Init
For reference, the generated files are stored in the Resources folder,
and the PDDL stuff should be moved there too.

The generator can now create the corresponding people and building
structures. For test purposes, you can also set what types of buildings
are currently present, what job people are and their current location.
As usual, this can all be done under Master Agent, and the file can be
generated only during runtime.