Integrity push
This should now be consistent and usable as a problem generator. Silly
me spent some time researching how to use the inspector better and
discarded his fancy inspector realising it was redundant.

Anyways, heres the push and the same ability to test it. If you can,
keep TaskPlannerProcess. The instructions are the same as the other
posts and I created some form of goal system with flexibility for other
types of goals if we need them.